+++ app/Http/Controllers/View/AuthenticationController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\View;
 
+use App\HttpOptionResponse\Auth\MigrateBudgetPro;
 use App\HttpOptionResponse\Auth\PermittedResourceType;
 use App\HttpOptionResponse\Auth\PermittedResourceTypeResource;
 use App\HttpOptionResponse\Auth\PermittedResourceTypeResources;
@@ -183,6 +184,15 @@ class AuthenticationController extends \Illuminate\Routing\Controller
         return response()->json($collection);
     }
 
+    public function optionsMigrateBudgetProRequestDelete(): Http\JsonResponse
+    {
+        $user = auth()->guard('api')->user();
+
+        $response = new MigrateBudgetPro(['view'=> $user !== null, 'manage'=> $user !== null]);
+
+        return $response->create()->response();
+    }
+
     public function optionsPermittedResourceType(): Http\JsonResponse
     {
         $user = auth()->guard('api')->user();
@@ -200,11 +210,7 @@ class AuthenticationController extends \Illuminate\Routing\Controller
     {
         $user = auth()->guard('api')->user();
 
-        if ($user === null) {
-            return Response::authenticationRequired();
-        }
-
-        $response = new PermittedResourceTypeResource(['view'=> true, 'manage'=> true]);
+        $response = new PermittedResourceTypeResource(['view'=> $user !== null, 'manage'=> $user !== null]);
 
         return $response->create()->response();
     }